refactor: rename variables for consistency and improve file structure
