+++ playground/src/context/layout.js
@@ -1,6 +1,6 @@
 import React from "react"
-import MainLayout from "@layout/MainLayout"
-import FullPageLayout from "@layout/FullPageLayout"
+import MainLayout from "@layout/main-layout"
+import FullPageLayout from "@layout/full-page-layout"
 
 const layouts = {
     main: MainLayout,
