+++ playground/src/features/schema/visualizer.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Alert} from "antd";
-import V from "@pkg/visualizer/index";
+import V from "@pkg/visualizer";
 import {useShapeStore} from "@state/shape";
 
 function Visualizer() {
