+++ playground/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './_assets/less/index.less';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import Main from './main';
+import reportWebVitals from './report-web-vitals';
 import { LoadWasm } from './wasm';
 
 const rootElement = document.getElementById('root');
@@ -13,7 +13,7 @@ const root = ReactDOM.createRoot(rootElement);
 root.render(
     <React.StrictMode>
         <LoadWasm>
-            <App />
+            <Main />
         </LoadWasm>
     </React.StrictMode>
 );
