+++ playground/src/main.test.js
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import Main from './main';
 
 test('renders learn react link', () => {
-  render(<App />);
+  render(<Main />);
   const linkElement = screen.getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
 });
